as.integer cast to NUMERIC instead of INTEGER/INT
in MSSQL casting to NUMERIC casts to integer, and is actually more flexible/graceful, since it returns NA if unable to convert individual rows. Whereas casting to  INTEGER/INT results in the entire column becoming NA if a single row cannot be converted to integer.
Address issue https://github.com/tidyverse/dbplyr/issues/380

+++ R/backend-mssql.R
@@ -83,7 +83,7 @@
       as.POSIXct    = sql_try_cast("TIMESTAMP"),
       as.numeric    = sql_try_cast("FLOAT"),
       as.double     = sql_try_cast("FLOAT"),
-      as.integer    = sql_try_cast("INTEGER"),
+      as.integer    = sql_try_cast("NUMERIC"), # in MSSQL this converts to integer
       as.integer64  = sql_try_cast("BIGINT"),
       as.character  = sql_try_cast("VARCHAR(MAX)"),
       log           = sql_prefix("LOG"),
